refactor/#22: 조회 url 수정 및 오타 수정
