+++ src/app.module.ts
@@ -6,7 +6,7 @@ import { PrismaModule } from "./prisma/prisma.module";
 import { ItemsModule } from "./items/items.module";
 import { InventoryModule } from "./inventory/inventory.module";
 import { UsersModule } from "./users/users.module";
-import { CommentModule } from "./comment/comment.module";
+import { CommentModule } from "./comments/comments.module";
 import { PresentsModule } from "./presents/presents.module";
 import { CharactersModule } from "./characters/characters.module";
 import { PostsModule } from "./posts/posts.module";
